+++ src/compute/tests/integration_tests.rs
@@ -41,8 +41,8 @@ use risingwave_storage::memory::MemoryStateStore;
 use risingwave_storage::table::batch_table::storage_table::StorageTable;
 use risingwave_stream::error::StreamResult;
 use risingwave_stream::executor::monitor::StreamingMetrics;
-use risingwave_stream::executor::state_table::StateTable;
 use risingwave_stream::executor::state_table_handler::SourceStateTableHandler;
+use risingwave_stream::executor::table::state_table::StateTable;
 use risingwave_stream::executor::{
     ActorContext, Barrier, Executor, MaterializeExecutor, Message, PkIndices, SourceExecutor,
 };
