+++ src/stream/src/common/mod.rs
@@ -18,7 +18,7 @@ pub use infallible_expr::*;
 use risingwave_common::row::Row;
 use risingwave_storage::StateStore;
 
-use crate::executor::state_table::{RowStream, StateTable};
+use crate::executor::table::state_table::{RowStream, StateTable};
 use crate::executor::StreamExecutorResult;
 
 mod builder;
