+++ src/stream/src/executor/global_simple_agg.rs
@@ -22,7 +22,7 @@ use risingwave_storage::StateStore;
 use super::aggregation::{
     agg_call_filter_res, iter_table_storage, AggChangesInfo, AggStateStorage,
 };
-use super::state_table::StateTable;
+use super::table::state_table::StateTable;
 use super::*;
 use crate::error::StreamResult;
 use crate::executor::aggregation::{generate_agg_schema, AggCall, AggGroup};
