+++ src/stream/src/executor/lookup/tests.rs
@@ -25,7 +25,7 @@ use risingwave_storage::StateStore;
 
 use crate::executor::lookup::impl_::LookupExecutorParams;
 use crate::executor::lookup::LookupExecutor;
-use crate::executor::state_table::StateTable;
+use crate::executor::table::state_table::StateTable;
 use crate::executor::test_utils::*;
 use crate::executor::{
     Barrier, BoxedMessageStream, Executor, MaterializeExecutor, Message, PkIndices,
