+++ src/stream/src/executor/managed_state/dynamic_filter.rs
@@ -25,11 +25,10 @@ use risingwave_common::buffer::Bitmap;
 use risingwave_common::row::{CompactedRow, Row};
 use risingwave_common::types::{ScalarImpl, VIRTUAL_NODE_SIZE};
 use risingwave_common::util::epoch::EpochPair;
-use risingwave_storage::table::streaming_table::state_table::prefix_range_to_memcomparable;
 use risingwave_storage::StateStore;
 
 use crate::executor::error::StreamExecutorError;
-use crate::executor::state_table::StateTable;
+use crate::executor::table::state_table::{prefix_range_to_memcomparable, StateTable};
 use crate::executor::StreamExecutorResult;
 
 type ScalarRange = (Bound<ScalarImpl>, Bound<ScalarImpl>);
