+++ src/stream/src/executor/managed_state/join/mod.rs
@@ -40,7 +40,7 @@ use self::iter_utils::zip_by_order_key;
 use crate::cache::{cache_may_stale, EvictableHashMap, ExecutorCache, LruManagerRef};
 use crate::executor::error::StreamExecutorResult;
 use crate::executor::monitor::StreamingMetrics;
-use crate::executor::state_table::StateTable;
+use crate::executor::table::state_table::StateTable;
 use crate::task::ActorId;
 
 type DegreeType = u64;
