+++ src/stream/src/executor/mod.rs
@@ -76,8 +76,8 @@ mod simple;
 mod sink;
 mod sort;
 pub mod source;
-pub mod state_table;
 pub mod subtask;
+pub mod table;
 mod top_n;
 mod union;
 mod watermark_filter;
