+++ src/stream/src/executor/mview/test_utils.rs
@@ -20,7 +20,7 @@ use risingwave_common::util::sort_util::OrderType;
 use risingwave_storage::memory::MemoryStateStore;
 use risingwave_storage::table::batch_table::storage_table::StorageTable;
 
-use crate::executor::state_table::StateTable;
+use crate::executor::table::state_table::StateTable;
 
 pub async fn gen_basic_table(row_count: usize) -> StorageTable<MemoryStateStore> {
     let state_store = MemoryStateStore::new();
