+++ src/stream/src/executor/table/mod.rs
@@ -1,0 +1,6 @@
+pub mod state_table;
+
+#[cfg(test)]
+pub mod test_batch_table;
+#[cfg(test)]
+pub mod test_streaming_table;
