+++ src/stream/src/executor/table/test_streaming_table.rs
@@ -20,15 +20,14 @@ use risingwave_common::row::Row;
 use risingwave_common::types::DataType;
 use risingwave_common::util::epoch::EpochPair;
 use risingwave_common::util::sort_util::OrderType;
+use risingwave_storage::memory::MemoryStateStore;
+use risingwave_storage::table::DEFAULT_VNODE;
 
-use crate::error::StorageResult;
-use crate::memory::MemoryStateStore;
-use crate::table::streaming_table::state_table::StateTable;
-use crate::table::DEFAULT_VNODE;
+use crate::executor::table::state_table::StateTable;
 
 // test state table
 #[tokio::test]
-async fn test_state_table() -> StorageResult<()> {
+async fn test_state_table() {
     let state_store = MemoryStateStore::new();
     let column_descs = vec![
         ColumnDesc::unnamed(ColumnId::from(0), DataType::Int32),
@@ -148,12 +147,10 @@ async fn test_state_table() -> StorageResult<()> {
         .await
         .unwrap();
     assert_eq!(row4_delete, None);
-
-    Ok(())
 }
 
 #[tokio::test]
-async fn test_state_table_update_insert() -> StorageResult<()> {
+async fn test_state_table_update_insert() {
     let state_store = MemoryStateStore::new();
     let column_descs = vec![
         ColumnDesc::unnamed(ColumnId::from(0), DataType::Int32),
@@ -319,7 +316,6 @@ async fn test_state_table_update_insert() -> StorageResult<()> {
         .await
         .unwrap();
     assert_eq!(row1_commit, None);
-    Ok(())
 }
 
 #[tokio::test]
