+++ src/stream/src/executor/test_utils.rs
@@ -159,7 +159,7 @@ pub mod agg_executor {
 
     use crate::common::StateTableColumnMapping;
     use crate::executor::aggregation::{AggCall, AggStateStorage};
-    use crate::executor::state_table::StateTable;
+    use crate::executor::table::state_table::StateTable;
     use crate::executor::{
         ActorContextRef, BoxedExecutor, Executor, GlobalSimpleAggExecutor, PkIndices,
     };
@@ -327,7 +327,7 @@ pub mod top_n_executor {
     use risingwave_common::util::sort_util::OrderType;
     use risingwave_storage::memory::MemoryStateStore;
 
-    use crate::executor::state_table::StateTable;
+    use crate::executor::table::state_table::StateTable;
     pub async fn create_in_memory_state_table(
         data_types: &[DataType],
         order_types: &[OrderType],
