+++ src/stream/src/from_proto/global_simple_agg.rs
@@ -17,7 +17,7 @@
 use super::agg_common::{build_agg_call_from_prost, build_agg_state_storages_from_proto};
 use super::*;
 use crate::executor::aggregation::AggCall;
-use crate::executor::state_table::StateTable;
+use crate::executor::table::state_table::StateTable;
 use crate::executor::GlobalSimpleAggExecutor;
 
 pub struct GlobalSimpleAggExecutorBuilder;
